requirements.txt
fastapi==0.115.2
pydantic==2.8.2
python-multipart==0.0.9
python-dotenv==1.0.1
